+++ tests/testthat/test_AcqFunctionMulti.R
@@ -26,7 +26,7 @@ test_that("AcqFunctionMulti works for single-objective instances", {
 })
 
 test_that("AcqFunctionMulti works for multi-objective instances", {
-  skip_if_not_installed("emoa")
+  skip_if_not_installed("moocore")
   skip_if_not_installed("fastGHQuad")
   inst = MAKE_INST(OBJ_1D_2, PS_1D, trm("evals", n_evals = 5L))
   surrogate = SurrogateLearnerCollection$new(